lightly standardize ThingComps

+++ Source/communityframework/communityframework/Comps/ThingComps/CompWithCheapHashInterval.cs
@@ -15,8 +15,7 @@ namespace CF
     public abstract class CompWithCheapHashInterval : ThingComp
     {       
         private int hashOffset = 0;
-        public bool IsCheapIntervalTick(int interval) => (int)(Find.TickManager.TicksGame + hashOffset) % interval == 0;
-
+        public bool IsCheapIntervalTick(int interval) => (Find.TickManager.TicksGame + hashOffset) % interval == 0;
         public override void PostSpawnSetup(bool respawningAfterLoad)
         {
             base.PostSpawnSetup(respawningAfterLoad);
